packages/roboview/pyproject.toml
[tool.poetry]
name = "roboview"
version = "0.1.0"
description = ""
authors = ["marcdavidsut <Marc.Sutjipto@viadee.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
robotframework = "^7.2.2"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
pygments = "^2.19.1"
scikit-learn = "^1.6.1"
numpy = "^2.2.4"
pydantic = "^2.11.4"
starlette = "^0.46.0"
pydantic-settings = "^2.12.0"
coloredlogs = "^15.0.1"
robotframework-robocop = "^7.2.0"
httpx = "^0.28.1"
robotframework-browser = "^19.12.4"
robotframework-databaselibrary = "^2.4.1"
robotframework-crypto = "^0.4.2"
robotframework-seleniumlibrary = "^6.8.0"

[tool.poetry.group.dev.dependencies]
pyright = "^1.1.408"
pytest = "^9.0.2"
deptry = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
# Exlude the openapi_client directory from the static type checker
exclude = [
    ".venv",
    "tests"
  ]
